+++ apps/web/src/views/universalFarms/components/PositionItem/PositionInfo.tsx
@@ -3,18 +3,13 @@ import { useTheme } from '@pancakeswap/hooks'
 import { useTranslation } from '@pancakeswap/localization'
 import { Currency, CurrencyAmount, Token } from '@pancakeswap/swap-sdk-core'
 import { FeeTier, Flex, Row, Skeleton, Tag, Text, useMatchBreakpoints } from '@pancakeswap/uikit'
-import { formatBigInt, formatNumber } from '@pancakeswap/utils/formatBalance'
+import { formatNumber } from '@pancakeswap/utils/formatBalance'
 import { DoubleCurrencyLogo, FiatNumberDisplay } from '@pancakeswap/widgets-internal'
-import BigNumber from 'bignumber.js'
 import { RangeTag } from 'components/RangeTag'
-import useAccountActiveChain from 'hooks/useAccountActiveChain'
-import { useCakePrice } from 'hooks/useCakePrice'
 import React, { memo, useMemo } from 'react'
-import { useStakedPositionsByUser } from 'state/farmsV3/hooks'
-import { useAccountV2PendingCakeReward } from 'state/farmsV4/state/accountPositions/hooks/useAccountV2PendingCakeReward'
-import { getUniversalBCakeWrapperForPool } from 'state/farmsV4/state/poolApr/fetcher'
 import { PoolInfo } from 'state/farmsV4/state/type'
 import styled from 'styled-components'
+import { useV2CakeEarning, useV3CakeEarning } from 'views/universalFarms/hooks/useCakeEarning'
 import { PoolApyButton } from '../PoolApyButton'
 
 export type PositionInfoProps = {
@@ -34,48 +29,6 @@ export type PositionInfoProps = {
   amount1?: CurrencyAmount<Token>
   pool?: PoolInfo | null
   detailMode?: boolean
-}
-
-const useV2CakeEarning = (pool: PoolInfo | null | undefined) => {
-  const { account } = useAccountActiveChain()
-  const cakePrice = useCakePrice()
-  const { chainId, lpAddress } = pool || {}
-  const bCakeConfig = useMemo(() => {
-    return chainId && lpAddress
-      ? getUniversalBCakeWrapperForPool({ chainId, lpAddress, protocol: pool?.protocol })
-      : null
-  }, [chainId, lpAddress, pool?.protocol])
-  const { data: pendingCake } = useAccountV2PendingCakeReward(account, {
-    chainId,
-    lpAddress,
-    bCakeWrapperAddress: bCakeConfig?.bCakeWrapperAddress,
-  })
-  const earningsAmount = useMemo(() => +formatBigInt(BigInt(pendingCake ?? 0), 5), [pendingCake])
-  const earningsBusd = useMemo(() => {
-    return new BigNumber(earningsAmount ?? 0).times(cakePrice.toString()).toNumber()
-  }, [cakePrice, earningsAmount])
-
-  return {
-    earningsAmount,
-    earningsBusd,
-  }
-}
-
-const useV3CakeEarning = (tokenId?: bigint) => {
-  const stackedTokenId = useMemo(() => (tokenId ? [tokenId] : []), [tokenId])
-  const cakePrice = useCakePrice()
-  const {
-    tokenIdResults: [pendingCake],
-  } = useStakedPositionsByUser(stackedTokenId)
-  const earningsAmount = useMemo(() => +formatBigInt(pendingCake || 0n, 4), [pendingCake])
-  const earningsBusd = useMemo(() => {
-    return new BigNumber(earningsAmount).times(cakePrice.toString()).toNumber()
-  }, [cakePrice, earningsAmount])
-
-  return {
-    earningsAmount,
-    earningsBusd,
-  }
 }
 
 export const PositionInfo = memo(
@@ -191,7 +144,7 @@ const V2Earnings = ({ pool }: { pool: PoolInfo | null | undefined }) => {
 }
 
 const V3Earnings = ({ tokenId }: { tokenId?: bigint }) => {
-  const { earningsAmount, earningsBusd } = useV3CakeEarning(tokenId)
+  const { earningsAmount, earningsBusd } = useV3CakeEarning(tokenId ? [tokenId] : [])
   return <Earnings earningsAmount={earningsAmount} earningsBusd={earningsBusd} />
 }
